+++ fragnix/include/primitive-memops.h
@@ -2,20 +2,21 @@
 #define haskell_primitive_memops_h
 
 #include <stdlib.h>
+#include <stddef.h>
 #include <HsFFI.h>
 
-void hsprimitive_memcpy( void *dst, int doff, void *src, int soff, size_t len );
-void hsprimitive_memmove( void *dst, int doff, void *src, int soff, size_t len );
+void hsprimitive_memcpy( void *dst, ptrdiff_t doff, void *src, ptrdiff_t soff, size_t len );
+void hsprimitive_memmove( void *dst, ptrdiff_t doff, void *src, ptrdiff_t soff, size_t len );
 
-void hsprimitive_memset_Word8 (HsWord8 *, int, int, HsWord);
-void hsprimitive_memset_Word16 (HsWord16 *, int, int, HsWord);
-void hsprimitive_memset_Word32 (HsWord32 *, int, int, HsWord);
-void hsprimitive_memset_Word64 (HsWord64 *, int, int, HsWord64);
-void hsprimitive_memset_Word (HsWord *, int, int, HsWord);
-void hsprimitive_memset_Ptr (HsPtr *, int, int, HsPtr);
-void hsprimitive_memset_Float (HsFloat *, int, int, HsFloat);
-void hsprimitive_memset_Double (HsDouble *, int, int, HsDouble);
-void hsprimitive_memset_Char (HsChar *, int, int, HsChar);
+void hsprimitive_memset_Word8 (HsWord8 *, ptrdiff_t, size_t, HsWord);
+void hsprimitive_memset_Word16 (HsWord16 *, ptrdiff_t, size_t, HsWord);
+void hsprimitive_memset_Word32 (HsWord32 *, ptrdiff_t, size_t, HsWord);
+void hsprimitive_memset_Word64 (HsWord64 *, ptrdiff_t, size_t, HsWord64);
+void hsprimitive_memset_Word (HsWord *, ptrdiff_t, size_t, HsWord);
+void hsprimitive_memset_Ptr (HsPtr *, ptrdiff_t, size_t, HsPtr);
+void hsprimitive_memset_Float (HsFloat *, ptrdiff_t, size_t, HsFloat);
+void hsprimitive_memset_Double (HsDouble *, ptrdiff_t, size_t, HsDouble);
+void hsprimitive_memset_Char (HsChar *, ptrdiff_t, size_t, HsChar);
 
 #endif
